requirements.txt
# Core dependencies for RAG implementation
langchain>=0.1.0
langchain-community>=0.0.38
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1

# PDF processing
PyPDF2>=3.0.1
pdfplumber>=0.10.3

# Embeddings and vector stores
sentence-transformers>=2.2.2
chromadb>=0.4.18
faiss-cpu>=1.7.4

# OpenAI API
openai>=1.0.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Text processing
nltk>=3.8.1
scikit-learn>=1.3.0

# Web interfaces
streamlit>=1.28.0
gradio>=4.0.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
requests>=2.31.0

# Jupyter notebook
jupyter>=1.0.0
ipykernel>=6.23.0
notebook>=7.0.0

# Optional: for advanced PDF processing
pypdf>=3.17.0

# Optional: for better text chunking
unstructured>=0.10.0
unstructured[pdf]>=0.10.0
